From: Dennis Coyne <coyne@ligo.caltech.edu> To: Rolf Bork <rolf@ligo.caltech.edu> Cc: Alex Ivanov <aivanov@ligo.caltech.edu>,      Virginio Sannibale <vsanni@ligo.caltech.edu>,      Riccardo DeSalvo <desalvo_r@ligo.caltech.edu>,      David Ottaway <ottaway@ligo.mit.edu>,      Alberto Stochino <stochino@ligo.caltech.edu> Subject: Re: CDS support Parts/Attachments:    1 Shown     24 lines  Text    2   OK     9.8 KB     Application ----------------------------------------
Rolf,
Ah, that's too bad. We had a rather urgent request for either you or Alex
to do for us/HAM-SAS. We would like the control topology to change as
indicated in the attached. Basically we'd like to have 3 parallel
compensation paths for each of the 4 modal controllers, all with a common
offset. In addition, one of the controllers should have a gain just before
it (in order to independently set the gain on the error into an integrator
from the output gain of the integrator. Can either you or Alex implement
this change in the next day or two?
         Dennis


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@589 6dcd42c9-f523-4c6d-aada-af552506706e