util: update fe_load_burt script
Replaces burt.py, and changes the interface.  model name is now a
non-optional argument.  specifying dcuid is not required, as the
script now attempts to determine the dcuid from the model name.

Adds many more checks that the model name is correct, and that the
specified file is in the proper place, with proper name, etc.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@3729 6dcd42c9-f523-4c6d-aada-af552506706e