Closes #905. SDF Read/Write Precision Problems.
When high precision edits where made to a field, the results where not getting properly saved.  Writting 1e-12 to a field, accepting the change, and saving the SDF file would record the field as having a value of 0.  This was due to the way that the SDF system cached values, it would hold a text copy in its internal structure.  The value would be read from EPICS and printed into the text field with format string "%.10lf".  When the SDF was written to disk it would read the text field converting the value from text to double, loosing any data more than 10 places right of the decimal point.

The solution was to include a field that holds the value as a double.  The text field is used for display.  The new field is used to store the data at full precision.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4059 6dcd42c9-f523-4c6d-aada-af552506706e