Integrating the CA SDF build of the sequencer into the RCG build process.
Added a SDF_TABLE_CA.adl file that has channel access specific changes.  This should be merged with SDF_TABLE.adl so that two almost identical files do not need to be kept in sync.

Modified the RCG to accept a new parameter in the parser (casdf, 1 = the CA SDF), and the feCodeGen.pl and fmseq.pl files to build the database and sequencer.  The database has two channels added and one extended.  The sequencer is build with the USE_SYSTEM_TIME and CA_SDF defines.

This is not complete.  The generated .cmd file needs to be modified by hand at this point.  First the .ini file that gives the list of channels to modify needs to be included, and second the sequencer must not be started as this is not intended to run on a front end with rfm, dolphins, ... and the sequencer code expects that environment.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4069 6dcd42c9-f523-4c6d-aada-af552506706e