Adding UI support for #889 SDF monitoring of individual filter bits.  This also addresses #921 because it adds a control for each filter module to the MEDM display.
This adds a column of shell command buttons to the SDF_TABLE view.  The buttons are only shown when the associated row is a filter module.  The button opens up a modified FILTALH type display which shows the current SDF mask & the SWSTAT word together allowing the users to select which bits they want to monitor.  This is done via shell scripting as a double indirection is needed to go from the line on the SDF table to the switch channel and the SDF MASK channel.

This also adds in code to fmseq.pl to generate 2040 new channels.  1000 channels to hold the SDF mask, 1000 channels to signal bit operations (1 per mask), and 40 channels to link from the SDF table row to the proper mask channel.  The SDF system has a hard coded limit of 1000 filter modules, hence the 1000 units here.  This is added here, as these channels are really specific to the SDF monitoring, and to keep the records out of the real time models.

Adds src/epics/util/sdffiltmon.py  - This is used to launch medm with the correct parameters to show SDF_FILT_MONITOR.adl
Adds src/epics/util/SDF_FILT_MONITOR.adl  - This medm screen shows the current filter mask and allows the user to edit what is monitored.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4120 6dcd42c9-f523-4c6d-aada-af552506706e