Close #889 (individual filter module monitoring) #981 (Beckhoff/CA SDF not seeing updates to ENUM channels).
Adjusted the UI to better integrate individual monitoring of filters in a filter bank.  Moved the filter monitor button on top of the regular monitor button (visible when the current line is a filter bank).  Integrated changes to filter monitoring in with the change count for revert,accept,monitor of regular channels.

Found and fixed an issue where only the first data point received on an ENUM channel was being read.  After classification as number/string it was not fully converted to a string/numeric SDF type and was ignored.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4140 6dcd42c9-f523-4c6d-aada-af552506706e