Split the main producer thread, cleaned up epics variable export, added more metrics.
In investgating frame writer instability, the main producer thread (for the fw build) was
split into two threads.  This was to split the load of receiving, checksumming, and distributing data.

The EPICS interface was also cleaned up.  When EPICS variables were be exported there were listed as
pvValue[2].  This has been replaced with inlined getters/setters and symbolic constants for the various
EPICS PVs.

The set of EPICS PVs has been expanded to give creater insight to the internal workings of the application.



git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.0@4249 6dcd42c9-f523-4c6d-aada-af552506706e