plit the raw framer threads in two.  Expand cpu affinity settings.  Add more EPICS variables.
Split the raw framers into two components:
1. move the data from the circual buffer to the frame structure, also doing CRC check
2. take a given frame and write it to disk

This makes it easier for a raw frame writing thread to catch up if its IO went long.  It will not need t
o do all of the data processing required before the frame can be written.

This change increases the amount of RAM needed for buffers, as it adds 2 more frames to memory if doing
both full and science frames.  The size of the main circular buffer may need to be decreased.

Also added some defines to allow setting cpu affinities on more of the daqd threads.  At this point it d
id not appear helpful, so all affinity values are set to 0.

Added more EPICS variables to show the states of the data processing frame writer threads.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.0@4250 6dcd42c9-f523-4c6d-aada-af552506706e