Fixed changed the symmetricom driver to always populate the usec field.
When calculating time a 3 tuple is returned:

(seconds, useconds (10^-6), nanoseconds (10^-9))

When returning time from a symmetricom/spectricom card only the firstfields are filled:

(seconds, useconds, 0)

When returning time from the system clock (no cards) only the first and third fields are filled:

(seconds, 0, nanoseconds)

This commit reworks the system time to return usecs as well.

(seconds, useconds, nanoseconds)

Such that the nanoseconds is the number of nanoseconds past the last usec referenced.

This way things that only read the seconds and useconds always get a value.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4447 6dcd42c9-f523-4c6d-aada-af552506706e