Updates to producer_dc_zmq.cc which allow data to flow properly.
Fixes to gps handling.

Use the data stream as the clock, the zmq data transfer has more jitter than the mx, so sometimes transfers go long, sometimes short, allow for that.

This version is the first that could be arrange the data well enough for frames to be written.

Still needs some correctness testing.  Lots of printf/cout debugging code will need to be removed.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4550 6dcd42c9-f523-4c6d-aada-af552506706e