Updates to the symmetricom driver
The symmetricom-4.4 driver now builds cleanly on kernel 3.0.8-4.9.

A new set of interfaces has been added to a gps_offset to be set
from user space.  The goal is that a new year/leap second should
not require a new code release, simple an update to the offset and
a restart of models.

Adds a /proc/gps_offset entry and a internal kernel interface
'long ligo_get_gps_offset(void)' which is consumed by the spectracom driver
in the IOPs, allowing the IOP to receive the offset as configured in the
system.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4553 6dcd42c9-f523-4c6d-aada-af552506706e