Adding a program to calculate gps offsets
Added a python program (and tests) to be used to calculate
the proper gps offset to use with the symmetricom/gpstime driver.

The program needs access to a gps leap second file as provied by
the ietf-leap-seconds package.  Then it can be queried to determine
what the offset should be.

To calculate the offset for a spectricom card that does not get the year
in the IRIG-B signal

calc_gps_offset.py -y -l leap-seconds.list

To calculate the offset for a system with NO timing card:

calc_gps_offset.py -s -l leap-seconds.list

Optionally you can specify a day for which you want the offset to be
calculated (defaults to today).  It is specified with the '-t' option
with the date in the YYYY-MM-DD form, ie '-t 2017-1-1'.

The idea is that this command can be put in startup scripts to calculate
the offset that needs to be written to /proc/gps_offset.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4555 6dcd42c9-f523-4c6d-aada-af552506706e