Forward port 4565. Updating run_number_client to handle timeouts/connection errors properly
Added a test_run_number_client that can be used to call the run_number server with arbitrary values.
Added more verbose output (when flagged with -v) output to the run_number server.
Added a test mode to the run_number server to have it crash/quit on a request to enable testing of that failure mode.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4568 6dcd42c9-f523-4c6d-aada-af552506706e