Reworked dataviewer cmake logic to not fail if Xpm/Xt not installed.
If development files for libXpm or libXt are not installed the cmake
process would crash.  The detection variables are not defined at all
if they are not found.

Rework the tests, such that it works as it should, the cmake finishes
and dataviewer is not built if dependencies are not found.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.4@4574 6dcd42c9-f523-4c6d-aada-af552506706e