Adapted the build system so that libzmq is optional.
If built without zmq present the src/zmq_stream will not be built and the daqd builds will use a run_number_client that always returns a run number of 0.  This is in response to a request to build the daqd variants w/o having a hard zmq dependency.

The change has been done to both the CMake and autotools build systems.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.4@4611 6dcd42c9-f523-4c6d-aada-af552506706e