Update the start trend configuration, so that you cannot define the length of minute trends.
The configuration on the running system is:

start trend 600 60;

The last value (60) can be any multiple of 60 and defines how many minutes are in a
minute trend.

This option is being changed to:

start trend optionalint;

So in the production system it would be:

start trend 600;

Do not allow multi-minute trends.  It is too easy to misconfigure and end up with your minute trends averaging data longer than you thought.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4748 6dcd42c9-f523-4c6d-aada-af552506706e