Updated the help and options for zmq_fe.
zmq_fe requires IFO/SITE environment variables to be set or to have the
gdstp directory passed in.  Update the options for the gdstp directory,
to aid startup scripts and keep the options more consistent with other
send/receive programs by changing the gdstp dir option from -g to -d in
order to not conflict with the ix group number.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4761 6dcd42c9-f523-4c6d-aada-af552506706e