Change how daqd_shmem determines where to put dcu data in the producer.
Previously it had started at the beginning of the buffer and move
dcu by dcu.  Now, use data from the channel scatter gather maps to
set where each dcu starts.

Working this way should isolate the effect of arithmetic errors in
writing into the producer move buffer.

This is only used in the DAQD_SHMEM, using this in more of the producers
would require more extensive changes in the initialize_vmpic code +
accounting for where TP data starts.

This fixes data layout issues that showed up in daqd_shmem if an expected
dcu was marked as bad.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4764 6dcd42c9-f523-4c6d-aada-af552506706e