Updates to debug/metrics in zmq_rcv_ix_xmit
Fixed the min/max/mean for the endpoints received.
The min/max had been set to latch.  Now they are reset
each 1/16s cycle.  The mean needed to be cleared each
cycle as well.

Added some diagnostics and epics output to report on
the relative arrival time of the endpoints.

RECV_BIN_MISS, RECV_BIN_[0-9] are 5ms bins.  They count
the number of endpoints reporting in that 5ms offset from
the first endpoint that reports.  This can give a report
on how tightly packed the input data is.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/trunk@4797 6dcd42c9-f523-4c6d-aada-af552506706e