Checkin of work on zmq_rcv_ix_xmit.
Added a zmq_rcv_ix_xmit_delay which introduces a static delay before
sending to attempt to ride out times when messages from the FE computers
are bunched together.

Currently it is configured to have a 1s buffer of input and to delay 3
full cycles before sending data (it has a much larger buffer than is
needed).  The code has been converted to C++ to add some additional
checks for type safety, and some aids in debugging.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.5@4845 6dcd42c9-f523-4c6d-aada-af552506706e