Fixes to TP data for mx_stream/mx_rcv. Updates to building on gentoo with cmake
Fixed some depenancy issues that showed up on the gentoo build with cmake.

Added a new option '-2' to mx_stream to allow sending the size of the
DQ data block in the channel count field of the rmIpcStr structure.  This
is used by mx_rcv to determine how much data is DQ and how much is test point data.

git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.5@4874 6dcd42c9-f523-4c6d-aada-af552506706e