Removing file I/O from the edcu to test the impact on CRC errors.
Testing a correlation between file system load and CRC errors that has
only showed up on the edc model.  Commented out the crc check of the ini
file as it is broken on the edc (it updates the crc that is sent when ever
there is a change to the ini file, instead of just marking a change as being
present).


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/branch-3.5@4942 6dcd42c9-f523-4c6d-aada-af552506706e