Captures some test work of useing EPICS v4
This was an exploration in 2017 of using EPICS v4 to
stream data from the FE to the daqd.

This code used a customized build of epics v4 (using
cmake to build instead of the epics build system).

There is a fundimental issue, in that if each change in
the data set is not received, EPICS combines/overwrites
changes.

This checkin is to keep some of the structure and code used
in the tests.


git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/advLigoRTS/branches/experiment-on-epics-v4-streaming@4954 6dcd42c9-f523-4c6d-aada-af552506706e