Added an alternate (faster) crc calculation to mx_stream.
This is a crc that works 4 bytes at a time instead of 1, it was written
for the GDS code, then pulled into the daqd.

This is put in as one idea to mitigate CRC errors (missing data) seen
in production.  The thought being that a missing data error is due to
timing, network, or cpu slowdowns.  This is to make the cpu even less
of an issue for mx_stream.