Guard against bad mbufs in local_dc.
This handles the case where a model is not running at startup and we
create an mbuf for it.  The contents are uninitialized and undefined.
Simply setting an impossible cycle counter for all models (not just the
IOP) protects us from processing uninitialized mbufs.  The buffers will
not be seen as ready until a model runs and sets a proper cycle counter.