Fixes to mx_fe
When resyncing to the IOP, reset the counter for all models.
This helps mx_fe skip models that have not been started (ie
their mbuf is unitialized).

Corrected some index values.  The ixDataBlock dcuheaders where
indexed with the wrong value, and could cause a bad tpBlockSize
to be caculated if a required model was stopped (this happens when
the ii index != the db index) which can manifest itself with the
process aborting after seeing an overflow.