Created a 'multi' fe stream generator.
fe_stream_test is a single front end outputing a rmIpcStr to mbuf.

fe_multi_stream_test can be used to simulate fe systems in bulk, it
outputs to a dac_multi_cycle_t in mbuf, along with populating a
ini directory and master file.  This is targeted for daqd development
and testing.

This adds a fe_generator_support library which holds some common code
between the two generators.

This also adds another check to fe_stream_check, verifying the data crc.