Updates to the work queues and how they are used/passed.
This is to help properly track lifetimes of the work queues.

Put in a wrapper around pthread_create to allow specifying an arbitrary job to run in the guise of a std::function.  This keeps the flexibility (scheduler, stack size, ...) of pthreads but gives the us the ability to call more than just a void* (*)(void*).

With these changes the work queues have been moved out of object scope for both the producer and the daqd objects.  A few small static routines that were used to start worker threads in the producer/daqd are removed.