Closes #98.  Check for dcuids that are not normal and abort.
This check can be turned off by setting the parameter "allow_any_dcu" = "1" in the daqdrc.

This allows dcuid's in the channel list that are not in the TP group or the MYRICOM group.  (We may be able to restrict the TP group as well).

Also removed some old code that is specific to the internal edcu, which along with not being used anymore gives a more obscure error message.