Removing unneeded GPS leap second handling code.
The daqd has carried code for at least 9 years to set the GPS leap seconds in the frameH structure that is not used, too small, and the configs where never updated.

Remove the code.  Putting the correct leap second value into the FrameH structure is done by FrameCPP in the call to SetGTime.  There is no need to keep this old code around.

* For now turning the set gps-leaps into a no-op in the config.
* Removing the leap seconds field from the circular buffer
* Removing some member variables and a helper function from daqd.