Converting the daqd trender threads to be controlled by a thread_handler.
Move all the trender threads to being tracked by a thread handler and having a uniform stop signal.

Add a timed_get function to the circular buffer which will allow queries to the circualr buffers to time out.  This is used to allow the trender to see a stop request.