More work on #96.  Correctly set the trender stop flag when stopping the trender.
The stopping_now_ flag should be set to true to signal trender threads to stop.

Also move some stack based variable sized arrays onto the heap in std::vectors to remove a place that the stack can be overflowed.

Added a test to ensure the daqd stops when there is a jump in the data, and updated the live test daqdrc to start the trender thread to highlight issues with stopping.