Allow the standalone_edc to produce data early.
Allow a negative delay (defaults to -10ms).  This starts the data write early when negative.  This is done to mitigate the race between the streamers/local_dc and the edcu both
triggering off of the 16hz time cycles.  This makes the edcu work more like a fe in that it gets its data ready early.