Closes #158, make mbuf conditionally verbose.
Expose a new sysfs endpoint, kernel/mbuf/verbosity.  This contains an
integer value which defaults to 0.  When set to non-zero the mbuf driver
is put into verbose mode and outputs debug information on each connection
and command.  If regular (non-verbose) it is silent.

Errors in allocation and issues on startup still always issue printk calls.