Additional bounds checks in local_dc, dix_xmit, dix_recv.c
Removed an unused buffer in local_dc.c.  Updated the local_dc checks to be against the size of the output buffer, instead of the old 4mb/s limit.  This could be put back in, but needs to take into account a variable sized output buffer.

Added some checks to dix_xmit/recv to test for size before reading from/writing to the dolphin buffers.