Change awgtpman systemd service dependency on the kernel module service from Wants to Requisite
This makes awgtpman fail if the module service is not loaded, but does systemd will not attempt to start the module.

Previously if the module failed, awgtpman would eventually restart, triggering a restart of the whole model.  This would lead to cluttered logs full of errors from each restart.
This chang fixes that issue.