Work on #186 parameter sanitation for daqd and transports.
This does some more checks in the daqd producer to verify that dcu id/counts are sane as well as there being no overflows.

Updates to dix_xmit and to the daq_core files to make total models and fulldatablocksize unsigned.  This allows a check for size to skip if the send length is < 0 before writing to dolphin.