Fixing a bug in cps_recv, track the last viewed time.
Cps_recv had not been properly tracking the last viewed cycle, so every time it looked at the shared memory it saw what it thought was a new message and sent (and resent) the data that was present in the buffer was re-broadcasted (in this case written into the mbuf in same spot multiple times).

Added a test case to the message queue that was used in helping to narrow down issues.