Changed the PV type for the checksum PVs in cps_recv to double to match daqd.
Daqd exports all PVs as doubles, changing the channel list checksum and data checksum to be doubles to match.  If this is not done, then a negative value is shown if the high bit is set when viewing the cps_recv version of the checksum, where the daqd shows a large positive number.